refactor(ci): Build openapi before proposing PR (#462)
Approach of downloading the github pages artifact for proposing a PR to
the OSC leads to the delay of 1 change (only after next PR is merged the
changes are proposed to the osc with the name of the last PR). This
is caused by the fact that it takes some seconds for openapi to be
published and so the propagate job mostly uses the outdated one.